pyproject.toml
[project]
name = "cbdep"
version = "1.3.1"
description = "Generalized tool installer"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Couchbase Build Team", email = "build-team@couchbase.com" }
]
requires-python = ">=3.10"
dependencies = [
    "PyYAML",
    "distro",
    "requests",
    "packaging",
]
classifiers=[
    'Development Status :: 4 - Beta',
    'Environment :: Console',
    'Intended Audience :: Developers',
    'License :: OSI Approved',
    'Operating System :: MacOS :: MacOS X',
    'Operating System :: POSIX'
]

[project.scripts]
cbdep = "cbdep.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "/tests"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.2.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "deptry>=0.20.0",
    "mypy>=0.991",
    "ruff>=0.6.9"
]

[tool.mypy]
files = ["src/cbdep"]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
